requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
asgiref==3.9.1
async-timeout==5.0.1
attrs==24.3.0
azure-core==1.32.0
azure-storage-blob==12.25.1
bcrypt==4.3.0
blinker==1.9.0
capstone==6.0.0a3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
colored-traceback==0.4.2
cryptography==44.0.2
cycler==0.11.0
Django==5.2.4
Flask==3.1.0
fonttools==4.29.1
frozenlist==1.5.0
idna==3.10
intervaltree==3.1.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.3.2
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.5.1
multidict==6.1.0
mysql-connector==2.2.9
mysql-connector-python==9.0.0
numpy==1.26.4
packaging==21.3
pandas==2.2.3
paramiko==3.5.1
pdfminer==20191125
Pillow==9.0.1
plumbum==1.9.0
prettytable @ https://pypi.python.org/packages/source/P/PrettyTable/prettytable-0.7.2.tar.bz2#sha256=853c116513625c738dc3ce1aee148b5b5757a86727e67eff6502c7ca59d43c36
propcache==0.2.1
psutil==7.0.0
pwn==1.0
pwntools==4.14.1
pybind11==2.13.6
pycparser==2.22
pycryptodome==3.20.0
pyelftools==0.32
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
pymysql2==1.3.3
PyNaCl==1.5.0
pyparsing==3.0.7
pyperclip==1.9.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.1.1
pytz==2024.2
pywin32==310
requests==2.32.3
ROPGadget==7.6
rpyc==6.0.1
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.5.3
twilio==9.4.1
typing_extensions==4.12.2
tzdata==2024.2
unicorn==2.1.2
unix-ar==0.2.1
urllib3==2.3.0
wcwidth==0.2.5
Werkzeug==3.1.3
yarl==1.18.3
zstandard==0.23.0
flask
azure-storage-blob
python-dotenv
